Move TimeProvider
