+++ Tests/Testably.Abstractions.Testing.Tests/Notification/NotificationTests.cs
@@ -1,15 +1,15 @@
 ﻿using System.Threading;
 
-namespace Testably.Abstractions.Testing.Tests;
+namespace Testably.Abstractions.Testing.Tests.Notification;
 
 public class NotificationTests
 {
 	[Fact]
 	public void AwaitableCallback_Amount_ShouldOnlyReturnAfterNumberOfCallbacks()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		int receivedCount = 0;
-		Notification.IAwaitableCallback<TimeSpan> wait =
+		Testing.Notification.IAwaitableCallback<TimeSpan> wait =
 			timeSystem.On.ThreadSleep(t =>
 			{
 				if (t.TotalMilliseconds > 0)
@@ -35,9 +35,9 @@ public class NotificationTests
 	[Fact]
 	public void AwaitableCallback_Filter_ShouldOnlyUpdateAfterFilteredValue()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		int receivedCount = 0;
-		Notification.IAwaitableCallback<TimeSpan> wait =
+		Testing.Notification.IAwaitableCallback<TimeSpan> wait =
 			timeSystem.On.ThreadSleep(_ =>
 			{
 				receivedCount++;
@@ -60,7 +60,7 @@ public class NotificationTests
 	[Fact]
 	public void AwaitableCallback_Predicate_ShouldOnlyUpdateAfterFilteredValue()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		int receivedCount = 0;
 		ManualResetEventSlim ms = new();
 		timeSystem.On.ThreadSleep(_ =>
@@ -90,9 +90,9 @@ public class NotificationTests
 		ManualResetEventSlim ms = new();
 		try
 		{
-			TimeSystemMock timeSystem = new();
+			Testing.TimeSystemMock timeSystem = new();
 			bool isCalled = false;
-			Notification.IAwaitableCallback<TimeSpan> wait =
+			Testing.Notification.IAwaitableCallback<TimeSpan> wait =
 				timeSystem.On.ThreadSleep(_ =>
 				{
 					isCalled = true;
@@ -122,10 +122,10 @@ public class NotificationTests
 		int secondThreadMilliseconds = 42;
 		int firstThreadMilliseconds = secondThreadMilliseconds + 1;
 		ManualResetEventSlim ms = new();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		bool isCalledFromSecondThread = false;
 		ManualResetEventSlim listening = new();
-		Notification.IAwaitableCallback<TimeSpan> wait =
+		Testing.Notification.IAwaitableCallback<TimeSpan> wait =
 			timeSystem.On
 			   .ThreadSleep(t =>
 				{
@@ -160,9 +160,9 @@ public class NotificationTests
 	[Fact]
 	public void AwaitableCallback_TimeoutExpired_ShouldThrowTimeoutException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		bool isCalled = false;
-		Notification.IAwaitableCallback<TimeSpan> wait = timeSystem.On.ThreadSleep(_ =>
+		Testing.Notification.IAwaitableCallback<TimeSpan> wait = timeSystem.On.ThreadSleep(_ =>
 		{
 			isCalled = true;
 		});
@@ -186,7 +186,7 @@ public class NotificationTests
 	[AutoData]
 	public void Execute_ShouldBeExecutedBeforeWait(int milliseconds)
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		int receivedMilliseconds = -1;
 		bool isExecuted = false;
 
@@ -211,7 +211,7 @@ public class NotificationTests
 	public void Execute_WithReturnValue_ShouldBeExecutedAndReturnValue(
 		int milliseconds, string result)
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		int receivedMilliseconds = -1;
 		bool isExecuted = false;
