+++ Tests/Testably.Abstractions.Testing.Tests/TimeProvider/TimeProviderTests.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Testably.Abstractions.Testing.Tests.TestHelpers;
 
-namespace Testably.Abstractions.Testing.Tests;
+namespace Testably.Abstractions.Testing.Tests.TimeProvider;
 
 public class TimeProviderTests
 {
@@ -10,7 +10,7 @@ public class TimeProviderTests
 	public void Now_ShouldReturnCurrentDateTime()
 	{
 		DateTime begin = DateTime.UtcNow;
-		TimeSystemMock.ITimeProvider timeProvider = TimeProvider.Now();
+		Testing.TimeSystemMock.ITimeProvider timeProvider = Testing.TimeProvider.Now();
 		DateTime end = DateTime.UtcNow;
 
 		DateTime result1 = timeProvider.Read();
@@ -27,7 +27,7 @@ public class TimeProviderTests
 
 		Parallel.For(0, 100, _ =>
 		{
-			results.Add(TimeProvider.Random().Read());
+			results.Add(Testing.TimeProvider.Random().Read());
 		});
 
 		results.Should().OnlyHaveUniqueItems();
@@ -37,7 +37,7 @@ public class TimeProviderTests
 	public void Use_ShouldReturnFixedDateTime()
 	{
 		DateTime now = TimeTestHelper.GetRandomTime();
-		TimeSystemMock.ITimeProvider timeProvider = TimeProvider.Use(now);
+		Testing.TimeSystemMock.ITimeProvider timeProvider = Testing.TimeProvider.Use(now);
 
 		DateTime result1 = timeProvider.Read();
 		DateTime result2 = timeProvider.Read();
