+++ Tests/Testably.Abstractions.Testing.Tests/TimeSystemMock/CallbackHandlerTests.cs
@@ -1,16 +1,16 @@
 ﻿using System.Threading;
 using Testably.Abstractions.Testing.Tests.TestHelpers;
 
-namespace Testably.Abstractions.Testing.Tests;
+namespace Testably.Abstractions.Testing.Tests.TimeSystemMock;
 
-public class TimeSystemMockCallbackHandlerTests
+public class CallbackHandlerTests
 {
 	[Fact]
 	public void
 		OnDateTimeRead_DisposedCallback_ShouldNotBeCalled()
 	{
 		DateTime expectedTime = TimeTestHelper.GetRandomTime(DateTimeKind.Local);
-		TimeSystemMock timeSystem = new(expectedTime);
+		Testing.TimeSystemMock timeSystem = new(expectedTime);
 		DateTime? receivedTime = null;
 		IDisposable disposable = timeSystem.On.DateTimeRead(d => receivedTime = d);
 
@@ -25,7 +25,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnDateTimeRead_MultipleCallbacks_DisposeOne_ShouldCallOtherCallbacks()
 	{
 		DateTime expectedTime = TimeTestHelper.GetRandomTime(DateTimeKind.Local);
-		TimeSystemMock timeSystem = new(expectedTime);
+		Testing.TimeSystemMock timeSystem = new(expectedTime);
 		DateTime? receivedTime1 = null;
 		DateTime? receivedTime2 = null;
 
@@ -44,7 +44,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnDateTimeRead_MultipleCallbacks_ShouldAllBeCalled()
 	{
 		DateTime expectedTime = TimeTestHelper.GetRandomTime(DateTimeKind.Local);
-		TimeSystemMock timeSystem = new(expectedTime);
+		Testing.TimeSystemMock timeSystem = new(expectedTime);
 		DateTime? receivedTime1 = null;
 		DateTime? receivedTime2 = null;
 
@@ -65,7 +65,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnDateTimeRead_Today_ShouldExecuteCallbackWithCorrectParameter()
 	{
 		DateTime expectedTime = TimeTestHelper.GetRandomTime().Date;
-		TimeSystemMock timeSystem = new(expectedTime);
+		Testing.TimeSystemMock timeSystem = new(expectedTime);
 		DateTime? receivedTime = null;
 
 		using (timeSystem.On.DateTimeRead(d => receivedTime = d))
@@ -81,7 +81,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnDateTimeRead_UtcNow_ShouldExecuteCallbackWithCorrectParameter()
 	{
 		DateTime expectedTime = TimeTestHelper.GetRandomTime(DateTimeKind.Utc);
-		TimeSystemMock timeSystem = new(expectedTime);
+		Testing.TimeSystemMock timeSystem = new(expectedTime);
 		DateTime? receivedTime = null;
 
 		using (timeSystem.On.DateTimeRead(d => receivedTime = d))
@@ -97,7 +97,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnTaskDelay_DisposedCallback_ShouldNotBeCalled()
 	{
 		int millisecondsDelay = new Random().Next();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan? receivedDelay = null;
 		IDisposable disposable = timeSystem.On.TaskDelay(d => receivedDelay = d);
 
@@ -112,7 +112,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnTaskDelay_MultipleCallbacks_DisposeOne_ShouldCallOtherCallbacks()
 	{
 		TimeSpan expectedDelay = TimeTestHelper.GetRandomInterval();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan? receivedDelay1 = null;
 		TimeSpan? receivedDelay2 = null;
 
@@ -131,7 +131,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnTaskDelay_MultipleCallbacks_ShouldAllBeCalled()
 	{
 		TimeSpan expectedDelay = TimeTestHelper.GetRandomInterval();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan? receivedDelay1 = null;
 		TimeSpan? receivedDelay2 = null;
 
@@ -152,7 +152,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnTaskDelay_WithMillisecondsAndWithCancellationToken_ShouldExecuteCallbackWithCorrectParameter()
 	{
 		int millisecondsDelay = new Random().Next();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan receivedDelay = TimeSpan.Zero;
 
 		using (timeSystem.On.TaskDelay(d => receivedDelay = d))
@@ -168,7 +168,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnTaskDelay_WithMillisecondsAndWithoutCancellationToken_ShouldExecuteCallbackWithCorrectParameter()
 	{
 		int millisecondsDelay = new Random().Next();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan receivedDelay = TimeSpan.Zero;
 
 		using (timeSystem.On.TaskDelay(d => receivedDelay = d))
@@ -184,7 +184,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnTaskDelay_WithTimeSpanAndWithCancellationToken_ShouldExecuteCallbackWithCorrectParameter()
 	{
 		TimeSpan expectedDelay = TimeTestHelper.GetRandomInterval();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan receivedDelay = TimeSpan.Zero;
 
 		using (timeSystem.On.TaskDelay(d => receivedDelay = d))
@@ -200,7 +200,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnTaskDelay_WithTimeSpanAndWithoutCancellationToken_ShouldExecuteCallbackWithCorrectParameter()
 	{
 		TimeSpan expectedDelay = TimeTestHelper.GetRandomInterval();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan receivedDelay = TimeSpan.Zero;
 
 		using (timeSystem.On.TaskDelay(d => receivedDelay = d))
@@ -216,7 +216,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnThreadSleep_DisposedCallback_ShouldNotBeCalled()
 	{
 		int millisecondsTimeout = new Random().Next();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan? receivedTimeout = null;
 		IDisposable disposable = timeSystem.On.ThreadSleep(d => receivedTimeout = d);
 
@@ -231,7 +231,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnThreadSleep_MultipleCallbacks_DisposeOne_ShouldCallOtherCallbacks()
 	{
 		TimeSpan expectedTimeout = TimeTestHelper.GetRandomInterval();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan? receivedTimeout1 = null;
 		TimeSpan? receivedTimeout2 = null;
 
@@ -250,7 +250,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnThreadSleep_MultipleCallbacks_ShouldAllBeCalled()
 	{
 		TimeSpan expectedTimeout = TimeTestHelper.GetRandomInterval();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan? receivedTimeout1 = null;
 		TimeSpan? receivedTimeout2 = null;
 
@@ -271,7 +271,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnThreadSleep_WithMilliseconds_ShouldExecuteCallbackWithCorrectParameter()
 	{
 		int millisecondsTimeout = new Random().Next();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan receivedTimeout = TimeSpan.Zero;
 
 		using (timeSystem.On.ThreadSleep(d => receivedTimeout = d))
@@ -287,7 +287,7 @@ public class TimeSystemMockCallbackHandlerTests
 		OnThreadSleep_WithTimeSpan_ShouldExecuteCallbackWithCorrectParameter()
 	{
 		TimeSpan expectedTimeout = TimeTestHelper.GetRandomInterval();
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		TimeSpan receivedTimeout = TimeSpan.Zero;
 
 		using (timeSystem.On.ThreadSleep(d => receivedTimeout = d))
