+++ Tests/Testably.Abstractions.Testing.Tests/TimeSystemMock/TimeSystemMockTests.cs
@@ -1,14 +1,14 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-namespace Testably.Abstractions.Testing.Tests;
+namespace Testably.Abstractions.Testing.Tests.TimeSystemMock;
 
 public class TimeSystemMockTests
 {
 	[Fact]
 	public async Task Delay_Infinite_ShouldNotThrowException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		Exception? exception =
 			await Record.ExceptionAsync(() => timeSystem.Task.Delay(Timeout.Infinite));
 
@@ -18,7 +18,7 @@ public class TimeSystemMockTests
 	[Fact]
 	public async Task Delay_InfiniteTimeSpan_ShouldNotThrowException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		Exception? exception =
 			await Record.ExceptionAsync(()
 				=> timeSystem.Task.Delay(Timeout.InfiniteTimeSpan));
@@ -29,7 +29,7 @@ public class TimeSystemMockTests
 	[Fact]
 	public async Task Delay_LessThanInfinite_ShouldNotThrowException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		Exception? exception =
 			await Record.ExceptionAsync(() => timeSystem.Task.Delay(-2));
 
@@ -40,7 +40,7 @@ public class TimeSystemMockTests
 	[Fact]
 	public async Task Delay_LessThanInfiniteTimeSpan_ShouldNotThrowException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		Exception? exception =
 			await Record.ExceptionAsync(()
 				=> timeSystem.Task.Delay(TimeSpan.FromMilliseconds(-2)));
@@ -52,7 +52,7 @@ public class TimeSystemMockTests
 	[Fact]
 	public void Sleep_Infinite_ShouldNotThrowException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		Exception? exception =
 			Record.Exception(() => timeSystem.Thread.Sleep(Timeout.Infinite));
 
@@ -62,7 +62,7 @@ public class TimeSystemMockTests
 	[Fact]
 	public void Sleep_InfiniteTimeSpan_ShouldNotThrowException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		Exception? exception =
 			Record.Exception(() => timeSystem.Thread.Sleep(Timeout.InfiniteTimeSpan));
 
@@ -72,7 +72,7 @@ public class TimeSystemMockTests
 	[Fact]
 	public void Sleep_LessThanInfinite_ShouldNotThrowException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		Exception? exception =
 			Record.Exception(() => timeSystem.Thread.Sleep(-2));
 
@@ -82,7 +82,7 @@ public class TimeSystemMockTests
 	[Fact]
 	public void Sleep_LessThanInfiniteTimeSpan_ShouldNotThrowException()
 	{
-		TimeSystemMock timeSystem = new();
+		Testing.TimeSystemMock timeSystem = new();
 		Exception? exception =
 			Record.Exception(()
 				=> timeSystem.Thread.Sleep(TimeSpan.FromMilliseconds(-2)));
